+++ opencog/atomutils/TypeUtils.h
@@ -26,6 +26,7 @@
 
 #include <opencog/atoms/base/Handle.h>
 #include <opencog/atoms/base/types.h>
+#include <opencog/atoms/core/VariableList.h>
 
 namespace opencog
 {
@@ -178,6 +179,18 @@ Type type_intersection(Type lhs, Type rhs);
 std::set<Type> type_intersection(Type lhs, const std::set<Type>& rhs);
 std::set<Type> type_intersection(const std::set<Type>& lhs,
                                  const std::set<Type>& rhs);
+
+/**
+ * Generate a VariableList of the free variables of a given atom h.
+ */
+VariableListPtr gen_varlist(const Handle& h);
+
+/**
+ * Given an atom h and its variable declaration vardecl, turn the
+ * vardecl into a VariableList if not already, and if undefined,
+ * generate a VariableList of the free variables of h.
+ */
+VariableListPtr gen_varlist(const Handle& h, const Handle& vardecl);
 
 /** @}*/
 }
