+++ opencog/atomutils/Unify.cc
@@ -30,6 +30,7 @@
 #include <opencog/atoms/base/Node.h>
 #include <opencog/atoms/core/ScopeLink.h>
 #include <opencog/atomutils/FindUtils.h>
+#include <opencog/atoms/pattern/PatternUtils.h>
 
 namespace opencog {
 
@@ -353,9 +354,117 @@ Handle Unify::substitute(BindLinkPtr bl, const TypedSubstitution& ts)
 	//
 	// TODO: make sure that ts.second contains the declaration of all
 	// variables
-	Handle h = bl->substitute(strip_context(ts.first), ts.second);
+	Handle h = substitute(bl, strip_context(ts.first), ts.second);
 
+	// TODO move that back into the substitute function as to not
+	// create a BindLink with ill quotations in it in the meantime.
 	return Handle(consume_ill_quotations(BindLinkCast(h)));
+}
+
+// TODO: add unit test in UnifyUTest for it
+Handle Unify::substitute(BindLinkPtr bl, const HandleMap& var2val,
+                         Handle vardecl)
+{
+	// Perform substitution over the existing variable declaration, if
+	// no alternative is provided
+	if (!vardecl and bl->get_vardecl())
+		vardecl = substitute_vardecl(bl->get_vardecl(), var2val);
+
+	const Variables variables = bl->get_variables();
+
+	// Turn the map into a vector of new variable names/values
+	HandleSeq values = variables.make_values(var2val);
+
+	// Substituted BindLink outgoings
+	HandleSeq hs;
+
+	// Perform substitution over the pattern term, then remove
+	// constant clauses
+	Handle clauses = variables.substitute_nocheck(bl->get_body(), values);
+	hs.push_back(remove_constant_clauses(vardecl, clauses));
+
+	// Perform substitution over the rewrite term
+	hs.push_back(variables.substitute_nocheck(bl->get_implicand(), values));
+
+	// Filter vardecl
+	vardecl = filter_vardecl(vardecl, hs);
+
+	// Insert vardecl in hs if defined
+	if (vardecl)
+		hs.insert(hs.begin(), vardecl);
+
+	// Create the substituted BindLink
+	return classserver().factory(Handle(createLink(hs, bl->getType())));
+}
+
+Handle Unify::substitute_vardecl(const Handle& vardecl,
+                                 const HandleMap& var2val)
+{
+	if (not vardecl)
+		return Handle::UNDEFINED;
+
+	Type t = vardecl->getType();
+
+	// Base cases
+
+	if (t == VARIABLE_NODE) {
+		auto it = var2val.find(vardecl);
+		// Only substitute if the variable is substituted by another variable
+		if (it != var2val.end() and it->second->getType() == VARIABLE_NODE)
+			return it->second;
+		return Handle::UNDEFINED;
+	}
+
+	// Recursive cases
+
+	HandleSeq oset;
+
+	if (t == VARIABLE_LIST) {
+		for (const Handle& h : vardecl->getOutgoingSet()) {
+			Handle nh = substitute_vardecl(h, var2val);
+			if (nh)
+				oset.push_back(nh);
+		}
+		if (oset.empty())
+			return Handle::UNDEFINED;
+	}
+	else if (t == TYPED_VARIABLE_LINK) {
+		Handle new_var = substitute_vardecl(vardecl->getOutgoingAtom(0),
+		                                    var2val);
+		if (new_var) {
+			oset.push_back(new_var);
+			oset.push_back(vardecl->getOutgoingAtom(1));
+		} else return Handle::UNDEFINED;
+	}
+	else {
+		OC_ASSERT(false, "Not implemented");
+	}
+	return classserver().factory(Handle(createLink(oset, t)));
+}
+
+// TODO: for now it is assumed clauses are connected by an AndLink
+// only. To fix that one needs to generalize
+// PatternLink::unbundle_clauses to make it usable in that code too.
+Handle Unify::remove_constant_clauses(const Handle& vardecl,
+                                      const Handle& clauses)
+{
+	// Extract variables
+	VariableListPtr vl = gen_varlist(clauses, vardecl);
+	const OrderedHandleSet& vars = vl->get_variables().varset;
+
+	// Remove constant clauses
+	Type t = clauses->getType();
+	HandleSeq hs;
+	if (t == AND_LINK) {
+		for (const Handle& clause : clauses->getOutgoingSet()) {
+			if (not is_constant(vars, clause)) {
+				hs.push_back(clause);
+			}
+		}
+	} else if (not is_constant(vars, clauses)) {
+		return clauses;
+	}
+	return Handle(createLink(hs, AND_LINK));
 }
 
 Unify::SolutionSet Unify::operator()()
@@ -824,6 +933,16 @@ HandleMap strip_context(const Unify::HandleCHandleMap& hchm)
 	return result;
 }
 
+/**
+ * Generate a VariableList of the free variables of a given contextual
+ * atom ch.
+ */
+VariableListPtr gen_varlist(const Unify::CHandle& ch)
+{
+	OrderedHandleSet free_vars = ch.get_free_variables();
+	return createVariableList(HandleSeq(free_vars.begin(), free_vars.end()));
+}
+
 Handle Unify::type_intersection(const CHandle& lch, const CHandle& rch) const
 {
 	return type_intersection(lch.handle, rch.handle, lch.context, rch.context);
@@ -933,50 +1052,6 @@ bool Unify::inherit(const std::set<Type>& lhs, const std::set<Type>& rhs) const
 		if (not inherit(ty, rhs))
 			return false;
 	return true;
-}
-
-/**
- * Generate a VariableList of the free variables of a given atom h.
- */
-VariableListPtr gen_varlist(const Handle& h)
-{
-	OrderedHandleSet vars = get_free_variables(h);
-	return createVariableList(HandleSeq(vars.begin(), vars.end()));
-}
-Handle gen_vardecl(const Handle& h)
-{
-	return Handle(gen_varlist(h));
-}
-
-/**
- * Generate a VariableList of the free variables of a given contextual
- * atom ch.
- */
-VariableListPtr gen_varlist(const Unify::CHandle& ch)
-{
-	OrderedHandleSet free_vars = ch.get_free_variables();
-	return createVariableList(HandleSeq(free_vars.begin(), free_vars.end()));
-}
-
-/**
- * Given an atom h and its variable declaration vardecl, turn the
- * vardecl into a VariableList if not already, and if undefined,
- * generate a VariableList of the free variables of h.
- */
-VariableListPtr gen_varlist(const Handle& h, const Handle& vardecl)
-{
-	if (not vardecl)
-		return gen_varlist(h);
-	else {
-		Type vardecl_t = vardecl->getType();
-		if (vardecl_t == VARIABLE_LIST)
-			return VariableListCast(vardecl);
-		else {
-			OC_ASSERT(vardecl_t == VARIABLE_NODE
-			          or vardecl_t == TYPED_VARIABLE_LINK);
-			return createVariableList(vardecl);
-		}
-	}
 }
 
 Variables merge_variables(const Variables& lhs, const Variables& rhs)
